utilities_pkg/requirements.txt
click==6.7
click-log
pymysql>=0.6,<1.0.0
aiomysql
mysql-replication>=0.22
path
requests
beautifulsoup4
py-linq
aiobotocore
fs
pillow
pyexiv2
